+++ packages/hubspot-fields-js/CHANGELOG.md
@@ -2,10 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.4.0] - 2022-02-04
+- Abstraction of methods
+	- Now includes CLI command and webapck plugins within this single package
+	- Dependencies added for the CLI
+	- Package can be installed locally to use the CLI for any directory.
+
 ## [1.3.1] - 2022-11-01
 - Update Readme Links
 
-## [1.3.0] - 2022-11-01
+## [1.3.0] - 2022-01-11
 - FIX: Recent changes to the compilation assets, were prevent assets from being emitted to the HubSpotAutoUploadPlugin for upload.
     - Plugin will now correctly check if the asset is emitted
 - Field JS files will no only be transformed if asset was emitted, it is the first compile or a file in one of the `extraDirsToWatch` was modifed
@@ -14,7 +20,7 @@ All notable changes to this project will be documented in this file.
     - eg. `<i> [FieldsPlugin] FieldsJS is tranforming src/hubl/modules/test.module/fields.js`
 
 ## [1.2.0] - 2022-01-01
-- BREAKING CHANGE: Js Partials are now imported relative to src file, NOT from final dist file. Field JS tranformation using the src file and then updates webpack's emission functionlity to properly handle the upload.
+- **BREAKING CHANGE**: Js Partials are now imported relative to src file, NOT from final dist file. Field JS tranformation using the src file and then updates webpack's emission functionlity to properly handle the upload.
     - A special mention and massive thanks goes out to [BJ Szyjakowski](https://github.com/jazzyclimber) from Spingroup for his contribution and collaboration on this feature.
 
 ```javascript